refactor(repository-linting): extract ValidateMultipleTechnologies to validate.go
main.go shrinks from 334 lines to 17 (package doc + struct only); the
validation orchestrator and its fail-condition helpers move to a dedicated
validate.go. Pure file split — no behavior change.

Refs #143 (P3).
